command line interface
